Create searchEmojis
